### 2021-07-14 #3492 @whedon
[PRE REVIEW]: TidyTensor: Utilities for multidimensional arrays as named hierarchical structures
**Submitting author:** @oneilsh (<a href="http://orcid.org/0000-0001-6220-7080">Shawn T. O&#39;Neil</a>)
**Repository:** <a href="https://github.com/oneilsh/tidytensor" target ="_blank">https://github.com/oneilsh/tidytensor</a>
**Version:** v0.9.1
