### 2021-07-15 #3492 @oneilsh
Possible reviewers from the google sheet and their interests/programming language preferences: JohnCoene, neuroimaginador, agisga. Possibilities not listed in the sheet include contributors to R deep/machine learning and/or 'tidy' infrastructure, e.g. https://github.com/rstudio/keras/graphs/contributors and https://github.com/tidymodels/tidymodels/graphs/contributors. I don't know any of these folks personally or from other collaborations.
