+++ paper.bib
@@ -5,6 +5,7 @@
   volume={5},
   number={1},
   year={2017},
+  doi={10.5334/jors.148},
   publisher={Ubiquity Press}
 }
 
