### 2021-08-06 #3543 @oneilsh
Hi @taless474, quick question: I just noticed a couple of minor typos in the documentation. Should I fix them right away or not make any adjustments to the codebase while the review is ongoing?
