### 2021-08-22 #3543 @omid-s
Hi @oneilsh, there are two license files, repo's [license](https://github.com/oneilsh/tidytensor/blob/master/LICENSE) link goes to a file with two lines that sets copy right to you, though I can see another "[license.md](https://github.com/oneilsh/tidytensor/blob/master/LICENSE.md)" that has a brief MIT Licence description, I was wondering which one are you planning to keep?
