### 2021-08-23 #3543 @oneilsh
Hi @omid-s, I believe that's an artifact of R's restrictions on the LICENSE file not containing the actual text of the license, whereas LICENSE.md is used more commonly as an optional extra. This [gitlab issue](https://gitlab.com/gitlab-org/gitlab/-/issues/15053) describes some frustrations developers have had there with the duplicate.

LICENSE + LICENSE.md is how [`usethis::use_mit_license()`](https://usethis.r-lib.org/reference/licenses.html) set it up. If you think I should remove the extra `LICENSE.md` I can do so, just let me know. Thanks!
