### 2021-08-26 oneilsh/tidytensor#2 @omid-s
[JOSS Review]  .bib Reference Title Cases
This issue is part of the JOSS review going on in openjournals/joss-reviews#3543

Hi @oneilsh, I noticed some minor stuff in the references. Can you please change the titles in the [.bib file](https://github.com/oneilsh/tidytensor/blob/master/paper.bib) to adhere to the guideline [here](https://pandoc.org/MANUAL.html#capitalization-in-titles)?
For some packages like _xarray_, _garray_, _abind_ names the title should be surrounded by {}s to maintain their lower case.
Also, if you could put _hoyer2017xarray_, and, _baydin2018automatic_  in Title Case it'll be great.
