+++ paper.bib
@@ -1,5 +1,5 @@
 @article{hoyer2017xarray,
-  title={xarray: ND labeled arrays and datasets in Python},
+  title={{xarray: N-D Labeled Arrays and Datasets in Python}},
   author={Hoyer, Stephan and Hamman, Joe},
   journal={Journal of Open Research Software},
   volume={5},
@@ -10,8 +10,8 @@
 }
 
   @Manual{garray,
-    title = {garray: Generalized Array Arithmetic for Ragged Arrays with Named
-Margins},
+    title = {{garray: Generalized Array Arithmetic for Ragged Arrays with Named
+Margins}},
     author = {Qingsheng Huang},
     year = {2018},
     note = {R package version 1.1.2},
@@ -19,15 +19,15 @@ Margins},
   }
 
 @Manual{stars,
-    title = {stars: Spatiotemporal Arrays, Raster and Vector Data Cubes},
+    title = {{stars: Spatiotemporal Arrays, Raster and Vector Data Cubes}},
     author = {Edzer Pebesma},
     year = {2021},
     note = {R package version 0.5-3},
     url = {https://CRAN.R-project.org/package=stars},
   }
 
-    @Manual{abind,
-    title = {abind: Combine Multidimensional Arrays},
+@Manual{abind,
+    title = {{abind: Combine Multidimensional Arrays},
     author = {Tony Plate and Richard Heiberger},
     year = {2016},
     note = {R package version 1.4-5},
@@ -36,7 +36,7 @@ Margins},
 
 
   @Manual{rkeras,
-    title = {keras: R Interface to 'Keras'},
+    title = {{keras: R Interface to 'Keras'}},
     author = {JJ Allaire and François Chollet},
     year = {2021},
     note = {R package version 2.4.0},
@@ -44,7 +44,7 @@ Margins},
   }
 
   @Manual{rtensorflow,
-    title = {tensorflow: R Interface to 'TensorFlow'},
+    title = {{tensorflow: R Interface to 'TensorFlow'}},
     author = {JJ Allaire and Yuan Tang},
     year = {2021},
     note = {R package version 2.4.0},
@@ -53,7 +53,7 @@ Margins},
 
 
   @Manual{rtorch,
-    title = {torch: Tensors and Neural Networks with 'GPU' Acceleration},
+    title = {{torch: Tensors and Neural Networks with 'GPU' Acceleration}},
     author = {Daniel Falbel and Javier Luraschi},
     year = {2021},
     note = {R package version 0.4.0},
@@ -61,7 +61,7 @@ Margins},
   }
 
  @article{baydin2018automatic,
-  title={Automatic differentiation in machine learning: a survey},
+  title={{Automatic Differentiation in Machine Learning: A Survey}},
   author={Baydin, Atilim Gunes and Pearlmutter, Barak A and Radul, Alexey Andreyevich and Siskind, Jeffrey Mark},
   journal={Journal of machine learning research},
   volume={18},
