### 2021-08-29 #3543 @omid-s
> Hi @omid-s, thank you for the thorough review! I've made a branch with the fixes, here's a quick link to the compiled paper: https://github.com/oneilsh/tidytensor/suites/3607702673/artifacts/87088522 If it looks good I will merge it in and close the issue.

It's all looking Great @oneilsh :) Thanks for the updates.
