### 2021-09-03 oneilsh/tidytensor#4 @nhejazi
Missing DOIs in JOSS paper
Please update `paper.bib` for the JOSS paper to include DOIs for the two manuscripts cited, Baydin et al. (2018) in JMLR and Hoyer & Hamman (2017) in JOSS. This is an easily resolved blocker for the "Software Paper" component of the ongoing review in https://github.com/openjournals/joss-reviews/issues/3543.
