### 2021-09-03 oneilsh/tidytensor#5 @nhejazi
Visibility of community guidelines
As part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/3543, a review item relates to the inclusion of community guidelines for interested contributors. The currently available `CONTRIBUTING.md` does satisfy this requirement; however, based on some experience, I would recommend including a short section in the `README` file and linking to `CONTRIBUTING.md` from there (e.g., https://github.com/tlverse/origami#contributions). This will help to point interested contributors in the right direction in terms of the scope and style of their contributions to `tidytensor`. To be clear, this is not a blocker for the review but a suggestion.
