### 2021-09-03 #3543 @nhejazi
Hi @oneilsh, great work on `tidytensor`! My apologies for the delay in getting to this review. I've just gone through most of the package documentation and items in the review checklist. I opened two issues in the `tidytensor` repo as part of the review; happy to discuss their contents in those issues. I have a bit more work to do in going through the package functionality but expect to have the first pass of this review completed tomorrow.
