### 2021-09-03 04d15db
Added Contributing section to README.md (RE #5)
