### 2021-09-03 oneilsh/tidytensor#5 @oneilsh
Hi @nhejazi! I like this idea. How does this look?
