### 2021-09-03 oneilsh/tidytensor#4 @oneilsh
Hi @nhejazi, thanks for the comment, using DOIs is a good practice I agree, especially in our link-all-the-data world :) Two questions though: I think that the Hoyer et al. paper has a DOI, though perhaps it showed up in my updated .bib addressing the other reviewer, link to that updated compiled PDF is here: https://github.com/oneilsh/tidytensor/suites/3607702673/artifacts/87088522

For Bayden et al., I'm having trouble locating a proper DOI. I did find this one: https://dl.acm.org/doi/10.5555/3122009.3242010 but that doesn't look like a proper DOI as it doesn't resolve at https://doi.org. Perhaps I should add the https://dl.acm.org/doi/10.5555/3122009.3242010 link as a URL rather than DOI entry?

Thanks!
~Shawn
