### 2021-09-03 oneilsh/tidytensor#5 @nhejazi
Great, LGTM! I'll close this since it's resolved by 04d15db.
