### 2021-09-03 oneilsh/tidytensor#4 @nhejazi
Thanks for pointing out the updated paper draft @oneilsh, I'd missed that in my initial review. The reference to the Hoyer et al. paper LGTM. For the Bayden et al. paper, I'm similarly unable to track down a URL but I think it probably will do to just use the JMLR URL (https://jmlr.org/papers/v18/17-468.html) or the URL referenced in the bibTeX entry on the website.
