### 2021-09-03 89edbcf
adding URL instead of DOI for Bayden et al. (RE #4)
