+++ paper.bib
@@ -66,5 +66,6 @@ Margins}},
   journal={Journal of machine learning research},
   volume={18},
   year={2018},
-  publisher={Journal of Machine Learning Research}
+  publisher={Journal of Machine Learning Research},
+  url={http://jmlr.org/papers/v18/17-468.html}
 }
