### 2021-09-09 #3543 @oneilsh
@taless474 I am curious about continuing improvements to the package post-publication; is there any best practice I should follow when making updates? Is the JOSS publication tied to a particular version of the package perhaps, such that creating a new version for changes post-review would make it obvious that those updates weren't part of the review? (I'm thinking especially around continuing documentation improvements e.g. updates to README.md, which I wouldn't normally mint a new version number for, but could.)
