### 2021-10-07 4f453e2
updates to vignette, v1.0 designation and docs rebuild
