### 2021-10-07 #3543 @oneilsh
Ok, v1.0.0 created with minor doc improvements and DOI 10.5281/zenodo.5555276. Thank you again!
